added the colision functions to the game
